+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
   ![Issue linking](./images/ss-cl-issue-linking.png)
 
+- Adds icons to remotes in the *GitLens* view based on the remote service provider
 - Adds multi-cursor support to current line annotations &mdash; closes [#291](https://github.com/eamodio/vscode-gitlens/issues/291)
 - Adds support to toggle annotations for each file individually or for all files at once &mdash; closes [#289](https://github.com/eamodio/vscode-gitlens/issues/289)
   - Adds new controls the interactive settings editor (*Open Settings* from the Command Palette) to configure this new behavior